feat(rpc): remove v01
